src/micro-services/coordination-service/requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
kazoo
asyncio
asyncio-nats-client